[SERVER-1247] Create a common members utility Functions moved to members.js. Changes in docs.sh to generate documentation. [SERVER-1336] also fixed.
