+++ frontend/express/app.js
@@ -45,6 +45,7 @@ var versionInfo = require('./version.info'),
     languages = require('../../frontend/express/locale.conf'),
     render = require('../../api/utils/render.js'),
     rateLimit = require("express-rate-limit"),
+    membersUtility = require("./libs/members.js"),
     argon2 = require('argon2');
 
 
@@ -120,44 +121,14 @@ if (countlyConfig.web && countlyConfig.web.track === "all") {
 
 var countlyConfigOrig = JSON.parse(JSON.stringify(countlyConfig));
 
-/**
-* Check remote configuration
-**/
-function recheckConfigs() {
-    var checkUrl = "https://count.ly/configurations/ce/tracking";
-    if (COUNTLY_TYPE !== "777a2bf527a18e0fffe22fb5b3e322e68d9c07a6") {
-        checkUrl = "https://count.ly/configurations/ee/tracking";
-    }
-    request(checkUrl, function(error, response, body) {
-        if (typeof body === "string") {
-            try {
-                body = JSON.parse(body);
-            }
-            catch (ex) {
-                body = null;
-            }
-        }
-        if (body) {
-            if (countlyConfigOrig.web.use_intercom && typeof body.intercom !== "undefined") {
-                countlyConfig.web.use_intercom = body.intercom;
-            }
-            if (typeof countlyConfigOrig.web.track === "undefined" && typeof body.stats !== "undefined") {
-                if (body.stats) {
-                    countlyConfig.web.track = null;
-                }
-                else {
-                    countlyConfig.web.track = "none";
-                }
-            }
-        }
-    });
-}
-recheckConfigs();
 
 var countlyDb = plugins.dbConnection(countlyConfig);
 //reference for consistency between app and api processes
 common.db = countlyDb;
+membersUtility.db = countlyDb;
 
+//checking remote configuration
+membersUtility.recheckConfigs(countlyConfigOrig, countlyConfig);
 /**
 * Create sha1 hash string
 * @param {string} str - string to hash
@@ -654,39 +625,14 @@ app.get(countlyConfig.path + '/', function(req, res) {
     res.redirect(countlyConfig.path + '/login');
 });
 
-var getSessionTimeoutInMs = function(req) {
-    var myTimeoutValue = parseInt(plugins.getConfig("frontend", req.session && req.session.settings).session_timeout) * 1000 * 60;
-    //max value used by set timeout function
-    if (myTimeoutValue > 2147483647) {
-        myTimeoutValue = 1800000;
-    }//30 minutes
-    return myTimeoutValue;
-};
 var extendSession = function(req) {
-    req.session.expires = Date.now() + getSessionTimeoutInMs(req);
-    if (req.session.auth_token) {
-        var ChangeTime = getSessionTimeoutInMs(req);
-        if (ChangeTime > 0) {
-            authorize.extend_token({token: req.session.auth_token, db: countlyDb, extendTill: Date.now() + ChangeTime}, function(err) {
-                if (err) {
-                    console.log(err);
-                }
-            });
-        }
-        else { //changed to not expire
-            authorize.extend_token({token: req.session.auth_token, db: countlyDb, extendBy: 0}, function(err) {
-                if (err) {
-                    console.log(err);
-                }
-            });
-        }
-    }
+    membersUtility.extendSession(req);
 };
 var checkRequestForSession = function(req, res, next) {
     if (parseInt(plugins.getConfig("frontend", req.session && req.session.settings).session_timeout)) {
         if (req.session.uid) {
             if (Date.now() > req.session.expires) {
-                clearSession(req, res, next);
+                membersUtility.logout(req, res);
                 res.redirect(countlyConfig.path + '/login?message=logout.inactivity');
             }
             else {
@@ -716,7 +662,7 @@ app.get(countlyConfig.path + '/ping', function(req, res) {
 });
 
 app.get(countlyConfig.path + '/configs', function(req, res) {
-    recheckConfigs();
+    membersUtility.recheckConfigs(countlyConfigOrig, countlyConfig);
     res.send("Success");
 });
 
@@ -744,32 +690,6 @@ app.get(countlyConfig.path + '/session', function(req, res, next) {
 app.get(countlyConfig.path + '/dashboard', checkRequestForSession);
 app.post('*', checkRequestForSession);
 
-/**
-* clear session info for request
-* @param {object} req - request object
-* @param {object} res - response object
-* @param {function} next - callback for next middleware
-**/
-function clearSession(req, res, next) {
-    if (req.session) {
-        if (req.session.uid && req.session.email) {
-            plugins.callMethod("userLogout", {req: req, res: res, next: next, data: {uid: req.session.uid, email: req.session.email, query: req.query}});
-        }
-        req.session.uid = null;
-        req.session.gadm = null;
-        req.session.email = null;
-
-        if (req.session.auth_token) {
-            countlyDb.collection("auth_tokens").remove({_id: req.session.auth_token});
-            req.session.auth_token = null;
-        }
-        req.session.settings = null;
-        res.clearCookie('uid');
-        res.clearCookie('gadm');
-        req.session.destroy(function() {});
-    }
-}
-
 app.get(countlyConfig.path + '/logout', function(req, res) {
     if (req.query.message) {
         res.redirect(countlyConfig.path + '/login?message=' + req.query.message);
@@ -780,7 +700,7 @@ app.get(countlyConfig.path + '/logout', function(req, res) {
 });
 
 app.post(countlyConfig.path + '/logout', function(req, res, next) {
-    clearSession(req, res, next);
+    membersUtility.logout(req, res);
     if (req.query.message) {
         res.redirect(countlyConfig.path + '/login?message=' + req.query.message);
     }
@@ -999,15 +919,7 @@ app.get(countlyConfig.path + '/dashboard', function(req, res, next) {
                 }
             }
             else {
-                if (req.session) {
-                    req.session.uid = null;
-                    req.session.gadm = null;
-                    req.session.email = null;
-                    req.session.settings = null;
-                    res.clearCookie('uid');
-                    res.clearCookie('gadm');
-                    req.session.destroy(function() {});
-                }
+                membersUtility.clearReqAndRes(req, res);
                 res.redirect(countlyConfig.path + '/login');
             }
         });
@@ -1107,34 +1019,19 @@ app.get(countlyConfig.path + '/reset/:prid', function(req, res) {
     }
 });
 
-app.post(countlyConfig.path + '/reset', function(req, res, next) {
-    var result = validatePassword(req.body.password);
-
-    if (result === false) {
-        if (req.body.password && req.body.again && req.body.prid) {
-            req.body.prid += "";
-
-            argon2Hash(req.body.password).then(password => {
-                countlyDb.collection('password_reset').findOne({ prid: req.body.prid }, function(err, passwordReset) {
-                    countlyDb.collection('members').findAndModify({ _id: passwordReset.user_id }, {}, { '$set': { "password": password } }, function(err2, member) {
-                        member = member && member.ok ? member.value : null;
-                        plugins.callMethod("passwordReset", { req: req, res: res, next: next, data: member });
-                        req.flash('info', 'reset.result');
-                        res.redirect(countlyConfig.path + '/login');
-                    });
-
-                    countlyDb.collection('password_reset').remove({ prid: req.body.prid }, function() { });
-                });
-            }).catch(function() {
+app.post(countlyConfig.path + '/reset', function(req, res/*, next*/) {
+    membersUtility.reset(req, function(valid, member) {
+        if (valid === false) {
+            if (member) {
+                req.flash('info', 'reset.result');
+                res.redirect(countlyConfig.path + '/login');
+            }
+            else {
                 res.render('reset', { languages: languages, countlyFavicon: req.countly.favicon, countlyTitle: req.countly.title, countlyPage: req.countly.page, "csrf": req.csrfToken(), "prid": req.body.prid, "message": "", path: countlyConfig.path || "", cdn: countlyConfig.cdn || "", themeFiles: req.themeFiles, inject_template: req.template });
-            });
+            }
         }
         else {
-            res.render('reset', { languages: languages, countlyFavicon: req.countly.favicon, countlyTitle: req.countly.title, countlyPage: req.countly.page, "csrf": req.csrfToken(), "prid": req.body.prid, "message": "", path: countlyConfig.path || "", cdn: countlyConfig.cdn || "", themeFiles: req.themeFiles, inject_template: req.template });
-        }
-    }
-    else {
-        res.render('reset',
+            res.render('reset',
             {
                 languages: languages,
                 countlyFavicon: req.countly.favicon,
@@ -1149,22 +1046,16 @@ app.post(countlyConfig.path + '/reset', function(req, res, next) {
                 message: result,
                 password_min: plugins.getConfig("security").password_min
             });
-    }
+        }
+
+    });
 });
 
 app.post(countlyConfig.path + '/forgot', function(req, res, next) {
     if (req.body.email) {
-        req.body.email = (req.body.email + "").trim();
-        countlyDb.collection('members').findOne({"email": req.body.email}, function(err, member) {
+        membersUtility.forgot(req, function(member) {
             if (member) {
-                var timestamp = Math.round(new Date().getTime() / 1000),
-                    prid = sha512Hash(member.username + member.full_name, timestamp);
-                member.lang = member.lang || req.body.lang || "en";
-                countlyDb.collection('password_reset').insert({"prid": prid, "user_id": member._id, "timestamp": timestamp}, {safe: true}, function() {
-                    countlyMail.sendPasswordResetInfo(member, prid);
-                    plugins.callMethod("passwordRequest", {req: req, res: res, next: next, data: req.body});
-                    res.render('forgot', { languages: languages, countlyFavicon: req.countly.favicon, countlyTitle: req.countly.title, countlyPage: req.countly.page, "message": "forgot.result", "csrf": req.csrfToken(), path: countlyConfig.path || "", cdn: countlyConfig.cdn || "", themeFiles: req.themeFiles, inject_template: req.template});
-                });
+                res.render('forgot', { languages: languages, countlyFavicon: req.countly.favicon, countlyTitle: req.countly.title, countlyPage: req.countly.page, "message": "forgot.result", "csrf": req.csrfToken(), path: countlyConfig.path || "", cdn: countlyConfig.cdn || "", themeFiles: req.themeFiles, inject_template: req.template});
             }
             else {
                 res.render('forgot', { languages: languages, countlyFavicon: req.countly.favicon, countlyTitle: req.countly.title, countlyPage: req.countly.page, "message": "forgot.result", "csrf": req.csrfToken(), path: countlyConfig.path || "", cdn: countlyConfig.cdn || "", themeFiles: req.themeFiles, inject_template: req.template});
@@ -1177,231 +1068,41 @@ app.post(countlyConfig.path + '/forgot', function(req, res, next) {
 });
 
 app.post(countlyConfig.path + '/setup', function(req, res, next) {
-    countlyDb.collection('members').count(function(err, memberCount) {
-        if (!err && memberCount === 0) {
-            if (req.body.full_name && req.body.username && req.body.password && req.body.email) {
-                argon2Hash(req.body.password).then(password => {
-                    req.body.email = (req.body.email + "").trim();
-                    req.body.username = (req.body.username + "").trim();
-                    var doc = {"full_name": req.body.full_name, "username": req.body.username, "password": password, "email": req.body.email, "global_admin": true, created_at: Math.floor(((new Date()).getTime()) / 1000), password_changed: Math.floor(((new Date()).getTime()) / 1000)};
-                    if (req.body.lang) {
-                        doc.lang = req.body.lang;
-                    }
-                    countlyDb.collection('members').insert(doc, {safe: true}, function(err2, member) {
-                        member = member.ops;
-                        if (countlyConfig.web.use_intercom) {
-                            var options = {uri: "https://try.count.ly/s", method: "POST", timeout: 4E3, json: {email: req.body.email, full_name: req.body.full_name, v: COUNTLY_VERSION, t: COUNTLY_TYPE}};
-                            request(options, function(a, c, b) {
-                                a = {};
-                                a.api_key = md5Hash(member[0]._id + (new Date).getTime());
-                                b && (b.in_user_id && (a.in_user_id = b.in_user_id), b.in_user_hash && (a.in_user_hash = b.in_user_hash));
+    membersUtility.setup(req, function(err) {
+        if (!err) {
+            res.redirect(countlyConfig.path + '/dashboard');
 
-                                countlyDb.collection("members").update({_id: member[0]._id}, {$set: a}, function() {
-                                    plugins.callMethod("setup", {req: req, res: res, next: next, data: member[0]});
-                                    req.session.uid = member[0]._id;
-                                    req.session.gadm = !0;
-                                    req.session.email = member[0].email;
-                                    req.session.install = true;
-                                    authorize.save({
-                                        db: countlyDb,
-                                        multi: true,
-                                        owner: req.session.uid,
-                                        purpose: "LoggedInAuth",
-                                        ttl: getSessionTimeoutInMs(req),
-                                        callback: function(err3, token) {
-                                            if (err3) {
-                                                console.log(err3);
-                                            }
-                                            if (token) {
-                                                req.session.auth_token = token;
-                                            }
-                                            res.redirect(countlyConfig.path + '/dashboard');
-                                        }
-                                    });
-                                });
-                            });
-                        }
-                        else {
-                            var a = {};
-                            a.api_key = md5Hash(member[0]._id + (new Date).getTime());
-
-                            countlyDb.collection("members").update({_id: member[0]._id}, {$set: a}, function() {
-                                req.session.uid = member[0]._id;
-                                req.session.gadm = !0;
-                                req.session.email = member[0].email;
-                                req.session.install = true;
-
-                                authorize.save({
-                                    db: countlyDb,
-                                    multi: true,
-                                    owner: req.session.uid,
-                                    purpose: "LoggedInAuth",
-                                    ttl: getSessionTimeoutInMs(req),
-                                    callback: function(err3, token) {
-                                        if (err3) {
-                                            console.log(err3);
-                                        }
-                                        if (token) {
-                                            req.session.auth_token = token;
-                                        }
-                                        res.redirect(countlyConfig.path + '/dashboard');
-                                    }
-                                });
-                            });
-                        }
-                    });
-                }).catch(function() {
-                    res.redirect(countlyConfig.path + '/setup');
-                });
-            }
-            else {
-                res.redirect(countlyConfig.path + '/setup');
-            }
         }
-        else if (err) {
-            res.status(500).send('Server Error');
+        else if (err === "User exists") {
+            res.redirect(countlyConfig.path + '/login');
+        }
+        else if (err === "Wrong request parameters") {
+            res.redirect(countlyConfig.path + '/setup');
         }
         else {
-            res.redirect(countlyConfig.path + '/login');
+            res.status(500).send('Server Error');
         }
     });
 });
 
 app.post(countlyConfig.path + '/login', function(req, res, next) {
-    if (req.body.username && req.body.password) {
-        req.body.username = (req.body.username + "").trim();
-        verifyMemberArgon2Hash(req.body.username, req.body.password, (err, member) => {
-            if (member) {
-                if (member.locked) {
-                    plugins.callMethod("loginFailed", {req: req, res: res, next: next, data: req.body});
-                    res.redirect(countlyConfig.path + '/login?message=login.locked');
-                }
-                else {
-                    plugins.callMethod("loginSuccessful", {req: req, res: res, next: next, data: member});
-                    if (countlyConfig.web.use_intercom && member.global_admin) {
-                        countlyStats.getOverall(countlyDb, function(statsObj) {
-                            request({
-                                uri: "https://try.count.ly/s",
-                                method: "POST",
-                                timeout: 4E3,
-                                json: {
-                                    email: member.email,
-                                    full_name: member.full_name,
-                                    v: COUNTLY_VERSION,
-                                    t: COUNTLY_TYPE,
-                                    u: statsObj["total-users"],
-                                    e: statsObj["total-events"],
-                                    a: statsObj["total-apps"],
-                                    m: statsObj["total-msg-users"],
-                                    mc: statsObj["total-msg-created"],
-                                    ms: statsObj["total-msg-sent"]
-                                }
-                            }, function(a, c, b) {
-                                a = {};
-                                if (b) {
-                                    if (b.in_user_id && b.in_user_id !== member.in_user_id) {
-                                        a.in_user_id = b.in_user_id;
-                                    }
-                                    if (b.in_user_hash && b.in_user_hash !== member.in_user_hash) {
-                                        a.in_user_hash = b.in_user_hash;
-                                    }
-                                }
-                                Object.keys(a).length && countlyDb.collection("members").update({_id: member._id}, {$set: a}, function() {});
-                            });
-                        });
-                    }
-                    if (!countlyConfig.web.track || countlyConfig.web.track === "GA" && member.global_admin || countlyConfig.web.track === "noneGA" && !member.global_admin) {
-                        countlyStats.getUser(countlyDb, member, function(statsObj) {
-                            var custom = {
-                                apps: (member.user_of) ? member.user_of.length : 0,
-                                platforms: {"$addToSet": statsObj["total-platforms"]},
-                                events: statsObj["total-events"],
-                                pushes: statsObj["total-msg-sent"],
-                                crashes: statsObj["total-crash-groups"],
-                                users: statsObj["total-users"]
-                            };
-                            var date = new Date();
-                            request({
-                                uri: "https://stats.count.ly/i",
-                                method: "GET",
-                                timeout: 4E3,
-                                qs: {
-                                    device_id: member.email,
-                                    app_key: "386012020c7bf7fcb2f1edf215f1801d6146913f",
-                                    timestamp: Math.round(date.getTime() / 1000),
-                                    hour: date.getHours(),
-                                    dow: date.getDay(),
-                                    user_details: JSON.stringify(
-                                        {
-                                            custom: custom
-                                        }
-                                    )
-                                }
-                            }, function() {});
-                        });
-                    }
-                    req.session.regenerate(function() {
-                        // will have a new session here
-                        req.session.uid = member._id;
-                        req.session.gadm = (member.global_admin === true);
-                        req.session.email = member.email;
-                        req.session.settings = member.settings;
-
-                        var update = {last_login: Math.round(new Date().getTime() / 1000)};
-                        if (typeof member.password_changed === "undefined") {
-                            update.password_changed = Math.round(new Date().getTime() / 1000);
-                        }
-                        if (req.body.lang && req.body.lang !== member.lang) {
-                            update.lang = req.body.lang;
-                        }
-                        if (Object.keys(update).length) {
-                            countlyDb.collection('members').update({_id: member._id}, {$set: update}, function() {});
-                        }
-                        if (parseInt(plugins.getConfig("frontend", member.settings).session_timeout)) {
-                            req.session.expires = Date.now() + parseInt(plugins.getConfig("frontend", member.settings).session_timeout) * 1000 * 60;
-                        }
-                        if (member.upgrade) {
-                            res.set({
-                                'Cache-Control': 'no-cache, private, no-store, must-revalidate, max-stale=0, post-check=0, pre-check=0',
-                                'Expires': '0',
-                                'Pragma': 'no-cache'
-                            });
-                        }
-                        //create token
-                        authorize.save({
-                            db: countlyDb,
-                            multi: true,
-                            owner: req.session.uid,
-                            ttl: getSessionTimeoutInMs(req),
-                            purpose: "LoggedInAuth",
-                            callback: function(err2, token) {
-
-                                if (err2) {
-                                    console.log(err2);
-                                }
-                                if (token) {
-                                    req.session.auth_token = token;
-                                }
-                                res.redirect(countlyConfig.path + '/dashboard');
-                                bruteforce.reset(req.body.username);
-
-                            }
-                        });
-
-
-                    });
-                }
+    membersUtility.login(req, res, function(member) {
+        if (member) {
+            if (member.locked) {
+                res.redirect(countlyConfig.path + '/login?message=login.locked');
             }
             else {
-                plugins.callMethod("loginFailed", {req: req, res: res, next: next, data: req.body});
-                bruteforce.fail(req.body.username);
-                res.redirect(countlyConfig.path + '/login?message=login.result');
+                res.redirect(countlyConfig.path + '/dashboard');
+                bruteforce.reset(req.body.username);
             }
-        });
-    }
-    else {
-        res.redirect(countlyConfig.path + '/login?message=login.result');
-    }
+        }
+        else {
+            res.redirect(countlyConfig.path + '/login?message=login.result');
+            if (req.body.username) {
+                bruteforce.fail(req.body.username);
+            }
+        }
+    });
 });
 
 app.get(countlyConfig.path + '/api-key', function(req, res, next) {
@@ -1544,169 +1245,20 @@ app.post(countlyConfig.path + '/apps/icon', function(req, res, next) {
     }
 });
 
-/**
-* Validate password based on configured settings
-* @param {string} password - password to validatePassword
-* @returns {vary} returns string if there is error, or false if everything is ok
-**/
-function validatePassword(password) {
-    if (password.length < plugins.getConfig("security").password_min) {
-        return "management-users.password.length";
-    }
-    if (plugins.getConfig("security").password_char && !/[A-Z]/.test(password)) {
-        return "management-users.password.has-char";
-    }
-    if (plugins.getConfig("security").password_number && !/\d/.test(password)) {
-        return "management-users.password.has-number";
-    }
-    if (plugins.getConfig("security").password_symbol && !/[^A-Za-z\d]/.test(password)) {
-        return "management-users.password.has-special";
-    }
-    return false;
-}
-
-/**
-* Removes all other active sessions for user
-* @param {string} userId - id of the user for which to remove sessions
-* @param {string} my_token - current auth token
-* @param {string} my_session - current session id
-**/
-function killOtherSessionsForUser(userId, my_token, my_session) {
-    countlyDb.collection('sessions_').find({"session": { $regex: userId }}).toArray(function(err, sessions) {
-        var delete_us = [];
-        for (var i = 0; i < sessions.length; i++) {
-            var parsed_data = "";
-            try {
-                parsed_data = JSON.parse(sessions[i].session);
-            }
-            catch (error) {
-                console.log(error);
-            }
-            if (sessions[i]._id !== my_session && parsed_data && parsed_data.uid === userId) {
-                delete_us.push(sessions[i]._id);
-            }
-        }
-        if (delete_us.length > 0) {
-            countlyDb.collection('sessions_').remove({'_id': {$in: delete_us}});
-        }
-    });
-    //delete other auth tokens with purpose:"LoggedInAuth"
-    countlyDb.collection('auth_tokens').remove({'owner': countlyDb.ObjectID(userId), 'purpose': "LoggedInAuth", '_id': {$ne: my_token}});
-}
-
 app.post(countlyConfig.path + '/user/settings', function(req, res, next) {
     if (!req.session.uid) {
         res.end();
         return false;
     }
-
-    var updatedUser = {};
-    if (req.body.username && req.body.api_key) {
-        if (req.body.api_key.length !== 32) {
-            res.send("user-settings.api-key-length");
-            return false;
+    membersUtility.settings(req, function(result, message) {
+        if (message) {
+            res.send(message);
         }
-        req.body.username = (req.body.username + "").trim();
-        updatedUser.username = req.body.username;
-        updatedUser.api_key = req.body.api_key;
-        if (req.body.lang) {
-            updatedUser.lang = req.body.lang;
+        else {
+            res.send(result);
         }
-        var change = JSON.parse(JSON.stringify(updatedUser));
-        countlyDb.collection('members').findOne({"_id": countlyDb.ObjectID(req.session.uid + "")}, function(err, member) {
-            if (err || !member) {
-                return res.send(false);
-            }
-            countlyDb.collection('members').findOne({username: req.body.username}, async function(err2, user) {
-                if (err) {
-                    return res.send(false);
-                }
-                member.change = change;
-                if ((user && user._id + "" !== req.session.uid + "") || err2) {
-                    res.send("username-exists");
-                }
-                else {
-                    if (req.body.old_pwd && req.body.old_pwd.length) {
-                        member.change.password = true;
-                        try {
-                            var newPassword_SHA5 = sha512Hash(req.body.new_pwd),
-                                newPassword_ARGON2 = await argon2Hash(req.body.new_pwd);
-                        }
-                        catch (ex) {
-                            return res.send(false);
-                        }
-
-                        let isUsedBefore = false;
-                        if (plugins.getConfig('security').password_rotation > 0) {
-                            // Check if used before
-                            const promises = [];
-                            const passwordHistory = member.password_history || [];
-
-                            for (let i = 0; i < passwordHistory.length; i++) {
-                                const oldPassword = passwordHistory[i];
-                                if (isArgon2Hash(oldPassword)) {
-                                    promises.push(verifyArgon2Hash(oldPassword, req.body.new_pwd));
-                                }
-                                else if (oldPassword === newPassword_SHA5) {
-                                    isUsedBefore = true;
-                                    break;
-                                }
-                            }
-
-                            if (!isUsedBefore && promises.length > 0) {
-                                try {
-                                    const promiseResults = await Promise.all(promises);
-                                    isUsedBefore = promiseResults.some(x => x === true);
-                                }
-                                catch (ex) {
-                                    return res.send(false);
-                                }
-                            }
-                        }
-
-                        if (req.body.new_pwd !== req.body.old_pwd && !isUsedBefore) {
-                            var passRes = validatePassword(req.body.new_pwd);
-                            if (passRes === false) {
-                                updatedUser.password = newPassword_ARGON2;
-                                updatedUser.password_changed = Math.round(new Date().getTime() / 1000);
-                                countlyDb.collection('members').update({"_id": countlyDb.ObjectID(req.session.uid + "")}, {'$set': updatedUser, $push: {password_history: {$each: [newPassword_ARGON2], $slice: -parseInt(plugins.getConfig('security').password_rotation)}}}, {safe: true}, function(err3, result) {
-                                    if (result && result.result && result.result.ok && result.result.nModified > 0 && !err3) {
-                                        killOtherSessionsForUser(req.session.uid, req.session.auth_token, req.sessionID);
-                                        plugins.callMethod("userSettings", {req: req, res: res, next: next, data: member});
-                                        res.send(updatedUser.password_changed + "");
-                                    }
-                                    else {
-                                        res.send("user-settings.old-password-not-match");
-                                    }
-                                });
-                            }
-                            else {
-                                res.send(passRes);
-                            }
-                        }
-                        else {
-                            res.send("user-settings.password-not-old");
-                        }
-                    }
-                    else {
-                        countlyDb.collection('members').update({"_id": countlyDb.ObjectID(req.session.uid + "")}, {'$set': updatedUser}, {safe: true}, function(err3, result) {
-                            if (result && !err3) {
-                                plugins.callMethod("userSettings", {req: req, res: res, next: next, data: member});
-                                res.send(true);
-                            }
-                            else {
-                                res.send(false);
-                            }
-                        });
-                    }
-                }
-            });
-        });
-    }
-    else {
-        res.send(false);
-        return false;
-    }
+        return result;
+    });
 });
 
 app.post(countlyConfig.path + '/user/settings/lang', function(req, res) {
@@ -1766,16 +1318,11 @@ app.post(countlyConfig.path + '/users/check/email', function(req, res) {
         res.send(false);
         return false;
     }
-    req.body.email = (req.body.email + "").trim();
-
-    countlyDb.collection('members').findOne({email: req.body.email}, function(err, member) {
-        if (member || err) {
-            res.send(false);
-        }
-        else {
-            res.send(true);
-        }
-    });
+    else {
+        membersUtility.checkEmail(req.body.email, function(result) {
+            res.send(result);
+        });
+    }
 });
 
 app.post(countlyConfig.path + '/users/check/username', function(req, res) {
@@ -1783,16 +1330,11 @@ app.post(countlyConfig.path + '/users/check/username', function(req, res) {
         res.send(false);
         return false;
     }
-    req.body.username = (req.body.username + "").trim();
-
-    countlyDb.collection('members').findOne({username: req.body.username}, function(err, member) {
-        if (member || err) {
-            res.send(false);
-        }
-        else {
-            res.send(true);
-        }
-    });
+    else {
+        membersUtility.checkUsername(req.body.username, function(result) {
+            res.send(result);
+        });
+    }
 });
 
 app.post(countlyConfig.path + '/graphnotes/create', function(req, res) {
@@ -1914,53 +1456,13 @@ app.get(countlyConfig.path + '/render', function(req, res) {
 });
 
 app.get(countlyConfig.path + '/login/token/:token', function(req, res) {
-    var token = req.params.token;
-    var pathUrl = req.url.replace(countlyConfig.path, "");
-    var urlParts = url.parse(pathUrl, true);
-    var fullPath = urlParts.pathname;
-
-    authorize.verify_return({
-        db: countlyDb,
-        token: token,
-        req_path: fullPath,
-        callback: function(valid) {
-            if (!valid) {
-                plugins.callMethod("tokenLoginFailed", {req: req, res: res, data: {token: token}});
-                return res.redirect(countlyConfig.path + '/login?message=login.result');
-            }
-
-            countlyDb.collection('members').findOne({"_id": countlyDb.ObjectID(valid)}, function(err, member) {
-                if (err || !member) {
-                    plugins.callMethod("tokenLoginFailed", {req: req, res: res, data: {token: token, token_owner: valid}});
-                    return res.redirect(countlyConfig.path + '/login?message=login.result');
-                }
-
-                req.session.regenerate(function() {
-                    req.session.uid = member._id;
-                    req.session.gadm = (member.global_admin === true);
-                    req.session.email = member.email;
-                    req.session.settings = member.settings;
-
-                    plugins.callMethod("tokenLoginSuccessful", {req: req, res: res, data: {username: member.username}});
-                    bruteforce.reset(member.username);
-                    authorize.save({
-                        db: countlyDb,
-                        multi: true,
-                        owner: req.session.uid,
-                        ttl: getSessionTimeoutInMs(req),
-                        purpose: "LoggedInAuthToken",
-                        callback: function(err2, authToken) {
-                            if (err2) {
-                                console.log(err2);
-                            }
-                            if (authToken) {
-                                req.session.auth_token = authToken;
-                            }
-                            res.redirect(countlyConfig.path + '/dashboard');
-                        }
-                    });
-                });
-            });
+    membersUtility.loginWithToken(req, function(member) {
+        if (member) {
+            bruteforce.reset(member.username);
+            res.redirect(countlyConfig.path + '/dashboard');
+        }
+        else {
+            res.redirect(countlyConfig.path + '/login?message=login.result');
         }
     });
 });
